fix(arch): move narrowing predicate query to its own module
`type_queries/flow.rs` sits at its size ratchet (2755 lines), so the new
`extract_predicate_signature_for_narrowing` query cannot live there
without tripping the arch guard. Move it to a dedicated
`type_queries/predicate_narrowing` module and re-export it; `flow.rs` is
restored byte-for-byte to its pre-change state. No behavior change.
